+++ src/terminal/commands/user/su.ts
@@ -1,7 +1,7 @@
 /**
  * @file Declares the su command.
  */
-import { Command } from "../commands";
+import { Command } from "../../commands";
 import { log, LogLevel } from "../../utils/log";
 
 export const suCommand = new Command({
