+++ src/terminal/commands/user/whoami.ts
@@ -1,7 +1,7 @@
 /**
  * @file Declares the whoami command.
  */
-import { Command } from "../commands";
+import { Command } from "../../commands";
 import { log, LogLevel } from "../../utils/log";
 
 export const whoamiCommand = new Command({
